requirements_mlops.txt
# Core ML Libraries
scikit-learn==1.7.2
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2

# MLOps Tools
mlflow==2.8.0
dvc==3.30.0
apache-airflow==2.7.3

# API Framework - VERSIONS COMPATIBLES AVEC AIRFLOW
fastapi==0.100.0
uvicorn[standard]==0.23.2
pydantic==2.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
SQLAlchemy==1.4.50

# Monitoring
evidently==0.7.15
prometheus-client==0.19.0
